pyproject.toml
[project]
name = "embody-papers-analyzer"
version = "1.0.0"
description = "Academic paper analysis system for embody-related research"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.1.3",
    "llama-cpp-python>=0.3.2",
    "loguru>=0.7.2",
    "openai>=1.57.0",
    "tiktoken>=0.8.0",
    "python-dotenv>=1.0.1",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
]
